only check for subs directly in package, not parent
Since this module is meant for testing exported subs, it should be
checking for subs directly in the test namespace and should ignore subs
in the parent namespace.

This will cause it to ignore subs in the UNIVERSAL namespace, such as
UNIVERSAL::import, which may exist if UNIVERSAL.pm is loaded, or on
newer perl versions.